requirements.txt
babel==2.17.0
cffi==1.17.1
colorama==0.4.6
contourpy==1.3.2
coverage==7.9.1
cryptography==45.0.4
cycler==0.12.1
distlib==0.3.9
filelock==3.18.0
fonttools==4.58.4
iniconfig==2.1.0
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
mysql-connector-python==9.3.0
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pycparser==2.22
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
tkcalendar==1.6.1
tzdata==2025.2
virtualenv==20.31.2
flask
openai>=1.0.0
pymupdf
python-dotenv
fpdf

